refactored everything, move objects between groups
The Project was a bit of a God object and the way groups
were setup was a bunch of bolt-on hacks. So I basically
rewrote everything so that it now makes (slightly more)
sense.

The Group/Item UI is now handled by the ObjectGroup and
ObjectGroupItem classes, and the canvas is now managed
and/or facade-d slightly by ObjectGroupItem instead of
just being attached to the Project.

Basically the project is the controller for these objects
and does significantly less DOM manipulation with that
being delegated to ObjectGroup and ObjectGroupItem.

All this nonsense was caused by implementing moving
objects to a different group. So that feature is also
snuck in this commit, along with some more general
drag-and-drop mechanisms.

I'm sure it all works, though. Pretty sure.

fixes #51
fixes #16